requirements.txt
streamlit==1.29.0
pyodbc==5.0.1
msal==1.25.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
plotly==5.17.0
streamlit-authenticator==0.2.3
redis==5.0.1
sqlalchemy==2.0.23
